backend/pyproject.toml
[tool.poetry]
name = "challenge_DigitalSys"
version = "0.1.0"
description = ""
authors = ["Léo Graciano <leonardoferreiragraciano@gmail.com>"]
readme = "README.md"
packages = [{include = "challenge_DigitalSys"}]

[tool.poetry.dependencies]
python = "^3.9"
django = ">=4.2"
djangorestframework = "^3.14.0"
django-cors-headers = "^3.13.0"
python-decouple = "^3.6"
django-redis = "^5.2.0"
django-filter = "^22.1"
celery = "^5.2.7"
django-celery-beat = "^2.4.0"
django-celery-results = "^2.4.0"
psycopg2-binary = "^2.9.5"
djangorestframework-simplejwt = "^5.2.2"


[tool.poetry.group.dev.dependencies]
autopep8 = "^1.7.0"
django-extensions = "^3.2.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"